inlab_study/middleware/robot/controller.py
import RPi.GPIO as GPIO

class RobotController:
    def __init__(self):
        self.setup_gpio()
        
    def setup_gpio(self):
        GPIO.setmode(GPIO.BCM)
        # Setup your pins
        
    def study_mode(self):
        # Study mode movements/sounds
        pass
        
    def training_mode(self):
        # Training mode movements/sounds
        pass