inlab_study/middleware/requirements.txt
flask
flask-cors
RPi.GPIO